app/requirements.txt
fastapi==0.94.0
uvicorn==0.21.0
pyjwt==2.6.0
passlib[bcrypt]==1.7.4
pymysql==1.0.2
cryptography==39.0.2
pydantic[email]==1.10.6
pytz==2022.7.1
python-jose==3.3.0
jinja2==3.1.2
python-multipart==0.0.6
